Adaptado teste de criacao de instancia Ad
